seguridad ruta y logout

+++ EvaSum3Back/src/main/java/com/everis/data/controllers/CategoriaController.java
@@ -3,6 +3,7 @@ package com.everis.data.controllers;
 import java.util.List;
 import java.util.Optional;
 
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,53 +24,75 @@ public class CategoriaController {
 	CategoriaService cService;
 
 	@RequestMapping("")
-	public String productos(Model model) {
+	public String productos(Model model, HttpSession session) {
 
-		List<Categoria> lista = cService.allCategorias();
-		model.addAttribute("categorias", lista);
-		return "categoria.jsp";
+		Integer registrado = (Integer) session.getAttribute("registrado");
+		if (registrado == 1) {
+			List<Categoria> lista = cService.allCategorias();
+			model.addAttribute("categorias", lista);
+			return "categoria.jsp";
+		}
+
+		return "redirect:/error-sesion";
 
 	}
 
 	@RequestMapping("/insertar")
-	public String guardar(@RequestParam(value = "nombre") String nombre, Model model) {
+	public String guardar(@RequestParam(value = "nombre") String nombre, Model model, HttpSession session) {
 
-		Categoria categoria = new Categoria();
-		categoria.setNombre(nombre);
+		Integer registrado = (Integer) session.getAttribute("registrado");
+		if (registrado == 1) {
 
-		categoria = cService.guardarCategoria(categoria);
+			Categoria categoria = new Categoria();
+			categoria.setNombre(nombre);
+			categoria = cService.guardarCategoria(categoria);
+			List<Categoria> lista = cService.allCategorias();
+			model.addAttribute("categorias", lista);
+			return "categoria.jsp";
+		}
 
-		List<Categoria> lista = cService.allCategorias();
-		model.addAttribute("categorias", lista);
-
-		return "categoria.jsp";
+		return "redirect:/error-sesion";
 	}
 
 	@RequestMapping("/eliminar")
-	public String eliminar(@RequestParam(value = "id") Long id, Model model) {
+	public String eliminar(@RequestParam(value = "id") Long id, Model model, HttpSession session) {
 
-		cService.deleteById(id);
+		Integer registrado = (Integer) session.getAttribute("registrado");
+		if (registrado == 1) {
+			cService.deleteById(id);
+			List<Categoria> lista = cService.allCategorias();
+			model.addAttribute("categorias", lista);
+			return "categoria.jsp";
 
-		List<Categoria> lista = cService.allCategorias();
-		model.addAttribute("categorias", lista);
-		return "categoria.jsp";
-
+		}
+		return "redirect:/error-sesion";
 	}
 
 	@RequestMapping("/editar")
-	public String editar(@RequestParam("id") Long id, Model model) {
-		Optional<Categoria> categoria = cService.findById(id);
-		model.addAttribute("categoria", categoria);
-		return "editarCategoria.jsp";
+	public String editar(@RequestParam("id") Long id, Model model, HttpSession session) {
+
+		Integer registrado = (Integer) session.getAttribute("registrado");
+		if (registrado == 1) {
+
+			Optional<Categoria> categoria = cService.findById(id);
+			model.addAttribute("categoria", categoria);
+			return "editarCategoria.jsp";
+		}
+
+		return "redirect:/error-sesion";
 	}
 
 	@RequestMapping("/actualizar")
-	public String actualizarCancion(@Valid @ModelAttribute("categoria") Categoria categoria
+	public String actualizarCancion(@Valid @ModelAttribute("categoria") Categoria categoria, HttpSession session) {
 
-	) {
-		Optional<Categoria> categoria1 = cService.findById(categoria.getId());
-		cService.actualizarCategoria(categoria);
-		return "redirect:/categoria";
+		Integer registrado = (Integer) session.getAttribute("registrado");
+		if (registrado == 1) {
+			Optional<Categoria> categoria1 = cService.findById(categoria.getId());
+			cService.actualizarCategoria(categoria);
+			return "redirect:/categoria";
+		}
+
+		return "redirect:/error-sesion";
 	}
 
 }
